#113/Separated playlist API into a file
